backend/engine/requirements.txt
backend==0.2.4.1
celery==5.4.0
Flask==3.0.3
flask_caching==2.3.0
flask_cors==4.0.1
flask_jwt_extended==4.6.0
flask_login==0.6.3
flask_marshmallow==1.2.1
flask_restful==0.3.10
flask_sqlalchemy==3.1.1
Jinja2==3.1.4
matplotlib==3.9.1
PyJWT==2.8.0
python-dotenv==1.0.1
python_bcrypt==0.3.2
redis==5.0.7
SQLAlchemy==2.0.31
Werkzeug==3.0.3
